error handling.

+++ scripts_seqModules/scripts_WGseq/allelic_ratios_WGseq.m
@@ -266,7 +266,7 @@ if (Make_figure == true)
 				count_vector     = [SNP_countA SNP_countT SNP_countG SNP_countC];
 				chr_lines_analyzed(chr_num) = chr_lines_analyzed(chr_num)+1;
 				chr_SNP_data_positions{chr_num}(chr_lines_analyzed(chr_num)) = SNP_coordinate;
-				chr_SNP_data_ratios   {chr_num}(chr_lines_analyzed(chr_num)) = max(count_vector)/sum(count_vector);
+				chr_SNP_data_ratios   {chr_num}(chr_lines_analyzed(chr_num)) = double(max(count_vector))/double(sum(count_vector));
 				chr_count             {chr_num}(chr_lines_analyzed(chr_num)) = sum(count_vector);
 				old_chr          = chr_num;
 			end;
